Split up tests into separate files
