+++ metaspace/graphql/src/tests/setupTestFramework.ts
@@ -6,5 +6,6 @@
 // const jasmineEnv = (jasmine as any).getEnv();
 // jasmineEnv.addReporter(failFast.init());
 
-// Mock out RabbitMQ as resolvers.js connects on import
+// Mock out libraries/files that connect to external services
 jest.mock('amqplib', () => ({connect: () => new Promise(() => null)}));
+jest.mock('../utils/smAPI');
